requirements.txt
tensorflow==2.6.0
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.19.5
opencv-python==4.5.3.56
scipy==1.7.1
lmdb==1.2.1
Pillow==8.3.1
tensorflow-addons==0.13.0
tqdm==4.62.2
tensorflow-datasets==4.4.0
pandas==1.3.3
